I updated the folder structure of the components, added the header and footer folder, now the components are better organized
