+++ app/layout.tsx
@@ -1,6 +1,6 @@
 import '@/styles/globals.css';
-import Footer from '@/components/Footer';
-import Header from '../components/Header';
+import Footer from '@/components/footer/Footer';
+import Header from '@/components/header/Header';
 import React from 'react';
 import { roboto, katibeh } from '@/components/fonts';
 import { getCategories } from '../database/database';
